requirements.txt
fastapi
uvicorn
langchain
chromadb
PyPDF2
python-docx
unstructured
groq
dotenv